+++ src/modules/mediacreator/MediaCreator.ts
@@ -3,7 +3,7 @@ import { type DIVEScene } from '../../scene/Scene';
 import { type DIVERenderer } from '../../renderer/Renderer';
 import type DIVEOrbitControls from '../../controls/OrbitControls';
 import { type Vector3Like } from 'three';
-import { ModuleRegistry } from '../registry/ModuleRegistry';
+import { ModuleRegistry } from '../_system/ModuleRegistry';
 
 declare global {
     interface ModuleClasses {
